Add nested schemas

+++ crates/control_plane/src/sql/sql.rs
@@ -1,14 +1,15 @@
+use crate::models::created_entity_response;
 use crate::sql::context::CustomContextProvider;
 use crate::sql::functions::parse_json::ParseJsonFunc;
 use crate::sql::planner::ExtendedSqlToRel;
-use arrow::array::{RecordBatch, UInt64Array};
-use arrow::datatypes::{DataType, Field, Schema as ArrowSchema};
+use arrow::array::RecordBatch;
 use datafusion::catalog::SchemaProvider;
 use datafusion::catalog_common::information_schema::InformationSchemaProvider;
 use datafusion::catalog_common::{ResolvedTableReference, TableReference};
 use datafusion::common::{plan_datafusion_err, Result};
 use datafusion::datasource::default_table_source::provider_as_source;
 use datafusion::execution::context::SessionContext;
+use datafusion::logical_expr::sqlparser::ast::Insert;
 use datafusion::logical_expr::{LogicalPlan, ScalarUDF};
 use datafusion::sql::parser::Statement as DFStatement;
 use datafusion::sql::sqlparser::ast::{ObjectName, Statement, CreateTable as CreateTableStatement, SetExpr, Query, Value, GroupByExpr};
@@ -62,13 +63,18 @@ impl SqlExecutor {
         warehouse_name: &String,
     ) -> Result<Vec<RecordBatch>> {
         if let Statement::CreateTable(create_table_statement) = statement {
-            let new_table_full_name = create_table_statement.name.to_string();
-            let _new_table_wh_id = create_table_statement.name.0[0].clone();
-            // Get database identifier from warehouse.db.db_1.table_name
-            let new_table_db =
-                &create_table_statement.name.0[1..create_table_statement.name.0.len() - 1];
-            let new_table_name = create_table_statement.name.0.last().unwrap().clone();
+            let mut new_table_full_name = create_table_statement.name.to_string();
+            let mut ident = create_table_statement.name.0;
+            if !new_table_full_name.starts_with(warehouse_name) {
+                new_table_full_name = format!("{}.{}", warehouse_name, new_table_full_name);
+                ident.insert(0, Ident::new(warehouse_name));
+            }
+            let _new_table_wh_id = ident[0].clone();
+            // Get database identifier from warehouse.db.schema.table_name
+            let new_table_db = &ident[1..ident.len() - 1];
+            let new_table_name = ident.last().unwrap().clone();
             let location = create_table_statement.location.clone();
+            let transient = create_table_statement.transient;
 
             // Replace the name of table that needs creation (for ex. "warehouse"."database"."table" -> "table")
             // And run the query - this will create an InMemory table
@@ -76,13 +82,18 @@ impl SqlExecutor {
                 name: ObjectName {
                     0: vec![new_table_name.clone()],
                 },
+                transient: false,
                 ..create_table_statement
             };
             let updated_query = modified_statement.to_string();
-            self.execute_with_custom_plan(&updated_query).await?;
 
             // Get schema of new table
             let plan = self.get_custom_logical_plan(&updated_query).await?;
+            self.ctx
+                .execute_logical_plan(plan.clone())
+                .await?
+                .collect()
+                .await?;
             let schema = Schema::builder()
                 .with_schema_id(0)
                 .with_identifier_field_ids(vec![])
@@ -97,7 +108,10 @@ impl SqlExecutor {
             let iceberg_catalog = catalog.as_any().downcast_ref::<IcebergCatalog>().unwrap();
             let rest_catalog = iceberg_catalog.catalog();
             let new_table_ident = Identifier::new(
-                &new_table_db.into_iter().map(|v| v.value.clone()).collect::<Vec<String>>(),
+                &new_table_db
+                    .into_iter()
+                    .map(|v| v.value.clone())
+                    .collect::<Vec<String>>(),
                 &new_table_name.value,
             );
             match rest_catalog.tabular_exists(&new_table_ident).await {
@@ -124,16 +138,20 @@ impl SqlExecutor {
                 .await
                 .unwrap();
 
-            // Copy data from InMemory table to created table
-            let insert_query =
-                format!("INSERT INTO {new_table_full_name} SELECT * FROM {new_table_name}");
-            let result = self.ctx.sql(&insert_query).await?.collect().await?;
+            // we don't need physical table for transient tables
+            if !transient {
+                // Copy data from InMemory table to created table
+                let insert_query =
+                    format!("INSERT INTO {new_table_full_name} SELECT * FROM {new_table_name}");
+                let result = self.execute_with_custom_plan(&insert_query).await?;
+                // self.ctx.sql(&insert_query).await?.collect().await?;
 
-            // Drop InMemory table
-            let drop_query = format!("DROP TABLE {new_table_name}");
-            self.ctx.sql(&drop_query).await?.collect().await?;
-
-            Ok(result)
+                // Drop InMemory table
+                let drop_query = format!("DROP TABLE {new_table_name}");
+                self.ctx.sql(&drop_query).await?.collect().await?;
+                return Ok(result);
+            }
+            Ok(created_entity_response())
         } else {
             Err(datafusion::error::DataFusionError::NotImplemented(
                 "Only CREATE TABLE statements are supported".to_string(),
@@ -141,15 +159,33 @@ impl SqlExecutor {
         }
     }
 
-    pub async fn create_schema(&self, name: SchemaName, warehouse_name: &str) -> Result<Vec<RecordBatch>> {
+    pub async fn create_schema(
+        &self,
+        name: SchemaName,
+        warehouse_name: &str,
+    ) -> Result<Vec<RecordBatch>> {
         match name {
             SchemaName::Simple(schema_name) => {
                 println!("Creating simple schema: {:?}", schema_name);
                 let catalog = self.ctx.catalog(warehouse_name).unwrap();
                 let iceberg_catalog = catalog.as_any().downcast_ref::<IcebergCatalog>().unwrap();
                 let rest_catalog = iceberg_catalog.catalog();
-                let namespace: Vec<String> = schema_name.0.iter().map(|ident| ident.value.clone()).collect();
-                rest_catalog.create_namespace(&Namespace::try_new(&namespace).unwrap(), None).await.unwrap();
+                let namespace_vec: Vec<String> = schema_name
+                    .0
+                    .iter()
+                    .map(|ident| ident.value.clone())
+                    .collect();
+                let single_layer_namespace = vec![namespace_vec.join(".")];
+                if rest_catalog
+                    .load_namespace(&Namespace::try_new(&single_layer_namespace).unwrap())
+                    .await
+                    .is_err()
+                {
+                    rest_catalog
+                        .create_namespace(&Namespace::try_new(&namespace_vec).unwrap(), None)
+                        .await
+                        .unwrap();
+                }
             }
             _ => {
                 return Err(datafusion::error::DataFusionError::NotImplemented(
@@ -157,33 +193,25 @@ impl SqlExecutor {
                 ));
             }
         }
-        let schema = Arc::new(ArrowSchema::new(vec![Field::new(
-            "count",
-            DataType::UInt64,
-            false,
-        )]));
-        Ok(vec![RecordBatch::try_new(
-            schema,
-            vec![Arc::new(UInt64Array::from(vec![0]))],
-        )?])
+        Ok(created_entity_response())
     }
 
     pub async fn get_custom_logical_plan(&self, query: &String) -> Result<LogicalPlan> {
         let state = self.ctx.state();
         let dialect = state.config().options().sql_parser.dialect.as_str();
-        let statement = state.sql_to_statement(query, dialect)?;
+        let mut statement = state.sql_to_statement(query, dialect)?;
+        statement = self.modify_statement_references(statement, "");
 
         if let DFStatement::Statement(s) = statement.clone() {
-            let references = state.resolve_table_references(&statement)?;
-            println!("References: {:?}", references);
             let mut ctx_provider = CustomContextProvider {
                 state: &state,
                 tables: HashMap::new(),
             };
 
+            let references = state.resolve_table_references(&statement)?;
+            println!("References: {:?}", references);
             for reference in references {
                 let resolved = self.resolve_table_ref(reference);
-
                 if let Entry::Vacant(v) = ctx_provider.tables.entry(resolved.to_string()) {
                     if let Ok(schema) = self.schema_for_ref(resolved.clone()) {
                         if let Some(table) = schema.table(&resolved.table).await? {
@@ -192,7 +220,6 @@ impl SqlExecutor {
                     }
                 }
             }
-
             for catalog in self.ctx.state().catalog_list().catalog_names() {
                 let provider = self.ctx.state().catalog_list().catalog(&catalog).unwrap();
                 for schema in provider.schema_names() {
@@ -210,7 +237,7 @@ impl SqlExecutor {
                     }
                 }
             }
-            println!("Tables: {:?}", ctx_provider.tables.keys());
+            // println!("Tables: {:?}", ctx_provider.tables.keys());
             let planner = ExtendedSqlToRel::new(&ctx_provider);
             planner.sql_statement_to_plan(*s)
         } else {
